requirements.txt
bottle
paste
requests
fuzzyparsers
psycopg2-binary
bcrypt
